[24.09.24 강내호] - 프리랜서 읽기 수정
