+++ WishD/src/main/webapp/WEB-INF/views/freelancer/freelancer_find.jsp
@@ -7,7 +7,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>프리랜서 찾기</title>
+    <title>WishD | 프리랜서 찾기</title>
     <link
             href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
             rel="stylesheet"
