+++ WishD/src/main/webapp/WEB-INF/views/freelancer/freelancer_reg.jsp
@@ -6,7 +6,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>프리랜서 등록</title>
+    <title>WishD | 프리랜서 등록</title>
     <link
             href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
             rel="stylesheet"
@@ -155,8 +155,12 @@
 								 style="color:white; height:200px; width:100%; resize:none;" 
 								 rows="20" 
 								 cols="20" 
-								 maxlength="5000"></textarea>
+								 maxlength="5000"
+								 required
+								 onblur="validateIntroduction()"></textarea>
+							<span id="introduction_error" class="text-danger" style="font-size: 12px;"></span> <!-- 경고 메시지 -->
 		    		</div>
+		    		
 	
 					<!-- 3-2) 링크 -->
 		    		<div class="col-md-8 my-1 px-5 p-1">
